+++ demo/components/DatePicker/index.tsx
@@ -2,6 +2,7 @@ import * as React from 'react';
 import { Story } from '../Story';
 import { BasicUsage } from './BasicUsage';
 import { ControlledDate } from './ControlledDate';
+import { DisabledDates } from './DisabledDates';
 import { MinMax } from './MinMax';
 import { Customization } from './Customization';
 import { Positioned } from './Positioned';
@@ -14,6 +15,7 @@ export const DatePicker = () => (
     <ControlledDate title="Контролирование только date" />
     <MinMax title="Ограничение даты min-max" />
     <Customization title="Кастомизация" />
+    <DisabledDates title="Disabled dates" />
     <Positioned title="Позиционирование" />
   </Story>
 );