+++ leda/components/DateTimePicker/types.ts
@@ -5,6 +5,8 @@ import {
 } from '../../src/DateTimeInput/types';
 
 export interface DateTimePickerProps extends DateTimeInputProps {
+  /* Dates that are disabled to be selected. Array of dates or dates ranges. */
+  disabledDates?: (Date | [Date, Date])[],
   /** Формат отображаемой и вводимой даты. По-умолчанию "dd.MM.yyyy hh:mm" */
   format?: string,
   /** Признак принудительного открытия календаря */