+++ leda/src/DateTimeInput/handlers/handleChange.ts
@@ -4,11 +4,16 @@ import { createMask, formatDateTime, stringToDate } from '../helpers';
 import { HandlersData } from '../types';
 import { ChangeEvent } from '../../MaskedInputBase/types';
 import { maskValue } from '../../MaskedInputBase/helpers';
+import { getIsDateDisabled } from '../../Calendar/helpers';
 import { COMPONENT_TYPES } from '../constants';
 
 export const createChangeHandler = ({ props, dispatch }: HandlersData) => (ev: ChangeEvent): void => {
   const {
-    format = 'dd.MM.yyyy', onChange, name, type = COMPONENT_TYPES.DATE_ONLY,
+    disabledDates,
+    format = 'dd.MM.yyyy',
+    name,
+    onChange,
+    type = COMPONENT_TYPES.DATE_ONLY,
   } = props;
 
   const mask = createMask(format, type);
@@ -22,8 +27,16 @@ export const createChangeHandler = ({ props, dispatch }: HandlersData) => (ev: C
   dispatch(setValue(newValue));
 
   // если в инпуте валидная дата - записываем в date, иначе - запиываем null
-  if (newDate && newDate.getDate()) dispatch(setDate(newDate));
-  else dispatch(setDate(null));
+  if (
+    newDate
+    && newDate.getDate()
+    && !getIsDateDisabled(newDate, disabledDates) // user should be unable to select disabled dates
+  ) {
+    dispatch(setDate(newDate));
+  } else {
+    dispatch(setDate(null));
+  }
+
   // контролируемый режим
   if (isFunction(onChange)) {
     onChange({